+++ package.json
@@ -23,5 +23,8 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
+  },
+  "dependencies": {
+    "cross-env": "^7.0.3"
   }
 }
